Trait framebuffer::pixel::Pixel

pub trait Pixel: Copy + Hash {
    fn composite_buffer(src: &[Self], dest: &mut [Self]);

    fn blend(self, other: Self) -> Self;

    fn weight_blend(origin: Self, other: Self, blend: f32) -> Self;
}

A pixel provides methods to blend with others.

Required methods

fn composite_buffer(src: &[Self], dest: &mut [Self])

Composites the src pixel slice to the dest pixel slice.

fn blend(self, other: Self) -> Self

blend with another pixel considering their extra channel.

fn weight_blend(origin: Self, other: Self, blend: f32) -> Self

Blend two pixels linearly with weights, as blend for origin and (1-blend) for other.

Loading content...