Crate framebuffer

Simple linux framebuffer abstraction. Examples can be found here.

Structs

Bitfield
Bitfield which is a part of VarScreeninfo.

FixScreeninfo
Struct as defined in /usr/include/linux/fb.h Note: type is a keyword in Rust and therefore has been changed to fb_type.

Framebuffer
Struct that should be used to work with the framebuffer. Direct usage of frame should not be necessary.

FramebufferError

VarScreeninfo
Struct as defined in /usr/include/linux/fb.h

Enums

FramebufferErrorKind
Kind of errors that can occur when dealing with the Framebuffer.

KdMode
Enum that can be used to set the current KdMode.