Crate framebuffer

This crate defines a Framebuffer structure, which is effectively a region of memory that is interpreted as a 2-D array of pixels.

Re-exports

pub use pixel::*;

Modules

pixel

Defines the Pixel trait as well as basic pixel formats, like RBG/RBGA.

Structs

Framebuffer

A framebuffer is a region of memory interpreted as a 2-D array of pixels. The memory buffer is a rectangular region with a width and height.

Functions

init

Initializes the final framebuffer based on VESA graphics mode information obtained during boot.