Module framebuffer::pixel

Defines the Pixel trait as well as basic pixel formats, like RBG/RBGA.

Structs

AlphaPixel

An Alpha Pixel is a pixel with an alpha channel

RGBPixel

An RGB Pixel is a pixel with no extra channel.

Traits

Pixel

A pixel provides methods to blend with others.