Struct framebuffer::Framebuffer

pub struct Framebuffer<P: Pixel> { /* fields omitted */ }

A framebuffer is a region of memory interpreted as a 2-D array of pixels. The memory buffer is a rectangular region with a width and height.

Implementations

impl<P: Pixel> Framebuffer<P>

pub fn new(
width: usize,
height: usize,
physical_address: Option<PhysicalAddress>
) -> Result<Framebuffer<P>, &'static str>

Creates a new framebuffer with rectangular dimensions of width * height, specified in number of pixels. If the physical_address is provided, the returned framebuffer will be final, i.e., mapped to the physical memory at that address, which is typically a hardware graphics device’s memory. If the physical_address is None, the returned framebuffer is a “virtual” one that renders to a randomly-allocated chunk of memory.

pub fn buffer_mut(&mut self) -> &mut BoxRefMut<MappedPages, [P]>

Returns a mutable reference to the mapped memory of the buffer.

pub fn buffer(&self) -> &BoxRefMut<MappedPages, [P]>

Returns a reference to the mapped memory of the buffer

pub fn get_size(&self) -> (usize, usize)

Returns the (width, height) of this framebuffer.

pub fn composite_buffer(&mut self, src: &[P], index: usize)

Composites src to the buffer starting from index.

pub fn draw_pixel(&mut self, coordinate: Coord, pixel: P)

Draw a pixel at the given 1coordinate1. The pixel will be blended with the existing pixel value at that coordinate in this framebuffer.

pub fn overwrite_pixel(&mut self, coordinate: Coord, pixel: P)

Overwites a pixel at the given coordinate in this framebuffer instead of blending it like draw_pixel.

pub fn get_pixel(&self, coordinate: Coord) -> Option<P>

Returns the pixel value at the given coordinate in this framebuffer.

pub fn fill(&mut self, pixel: P)

Fills (overwrites) the entire framebuffer with the given pixel value.

pub fn index_of(&self, coordinate: Coord) -> Option<usize>

Returns the index of the given coordinate in this framebuffer, if this framebuffer contains the coordinate within its bounds.

pub fn contains(&self, coordinate: Coord) -> bool

Checks if the given coordinate is within the framebuffer’s bounds. The coordinate is relative to the origin coordinate of (0, 0) being the top-left point of the framebuffer.

pub fn overlaps_with(
&mut self,
coordinate: Coord,
width: usize,
height: usize
) -> bool

Checks if a framebuffer overlaps with an area.

Arguments

coordinate: the top-left corner of the area relative to the origin(top-left point) of the framebuffer.

width: the width of the area in number of pixels.

height: the height of the area in number of pixels.